refactor: extract password recovery content to component
